feat(ui): ajustement dashboard administrateur
